+++ corpus/invalid_unicode.clj
@@ -1,0 +1,3 @@
+(ns invalid-unicode)
+
+\u12345